moved default credential selection to the new credentials option page and created a cred records combobox
