+++ mRemoteV1/UI/Forms/OptionsPages/ConnectionsPage.cs
@@ -1,7 +1,5 @@
 using System;
-using mRemoteNG.App;
 using mRemoteNG.Config;
-using mRemoteNG.Security.SymmetricEncryption;
 
 namespace mRemoteNG.UI.Forms.OptionsPages
 {
@@ -33,14 +31,6 @@ namespace mRemoteNG.UI.Forms.OptionsPages
             lblAutoSave1.Text = Language.strAutoSaveEvery;
             lblAutoSave2.Text = Language.strAutoSaveMins;
 
-            lblDefaultCredentials.Text = Language.strEmptyUsernamePasswordDomainFields;
-            radCredentialsNoInfo.Text = Language.strNoInformation;
-            radCredentialsWindows.Text = Language.strMyCurrentWindowsCreds;
-            radCredentialsCustom.Text = Language.strTheFollowing;
-            lblCredentialsUsername.Text = Language.strLabelUsername;
-            lblCredentialsPassword.Text = Language.strLabelPassword;
-            lblCredentialsDomain.Text = Language.strLabelDomain;
-
             lblClosingConnections.Text = Language.strLabelClosingConnections;
             radCloseWarnAll.Text = Language.strRadioCloseWarnAll;
             radCloseWarnMultiple.Text = Language.strRadioCloseWarnMultiple;
@@ -61,26 +51,6 @@ namespace mRemoteNG.UI.Forms.OptionsPages
             numRDPConTimeout.Value = Convert.ToDecimal(Settings.Default.ConRDPOverallConnectionTimeout);
 
             numAutoSave.Value = Convert.ToDecimal(Settings.Default.AutoSaveEveryMinutes);
-
-            // ReSharper disable once StringLiteralTypo
-            // ReSharper disable once SwitchStatementMissingSomeCases
-            switch (Settings.Default.EmptyCredentials)
-            {
-                case "noinfo":
-                    radCredentialsNoInfo.Checked = true;
-                    break;
-                case "windows":
-                    radCredentialsWindows.Checked = true;
-                    break;
-                case "custom":
-                    radCredentialsCustom.Checked = true;
-                    break;
-            }
-
-            txtCredentialsUsername.Text = Settings.Default.DefaultUsername;
-            var cryptographyProvider = new LegacyRijndaelCryptographyProvider();
-            txtCredentialsPassword.Text = cryptographyProvider.Decrypt(Settings.Default.DefaultPassword, Runtime.EncryptionKey);
-            txtCredentialsDomain.Text = Settings.Default.DefaultDomain;
 
             switch (Settings.Default.ConfirmCloseConnection)
             {
@@ -120,25 +90,6 @@ namespace mRemoteNG.UI.Forms.OptionsPages
                 FrmMain.Default.tmrAutoSave.Enabled = false;
             }
 
-            if (radCredentialsNoInfo.Checked)
-            {
-                // ReSharper disable once StringLiteralTypo
-                Settings.Default.EmptyCredentials = "noinfo";
-            }
-            else if (radCredentialsWindows.Checked)
-            {
-                Settings.Default.EmptyCredentials = "windows";
-            }
-            else if (radCredentialsCustom.Checked)
-            {
-                Settings.Default.EmptyCredentials = "custom";
-            }
-
-            Settings.Default.DefaultUsername = txtCredentialsUsername.Text;
-            var cryptographyProvider = new LegacyRijndaelCryptographyProvider();
-            Settings.Default.DefaultPassword = cryptographyProvider.Encrypt(txtCredentialsPassword.Text, Runtime.EncryptionKey);
-            Settings.Default.DefaultDomain = txtCredentialsDomain.Text;
-
             if (radCloseWarnAll.Checked)
             {
                 Settings.Default.ConfirmCloseConnection = (int) ConfirmCloseEnum.All;
@@ -159,14 +110,6 @@ namespace mRemoteNG.UI.Forms.OptionsPages
             Settings.Default.Save();
         }
 
-        private void radCredentialsCustom_CheckedChanged(object sender, EventArgs e)
-        {
-            lblCredentialsUsername.Enabled = radCredentialsCustom.Checked;
-            lblCredentialsPassword.Enabled = radCredentialsCustom.Checked;
-            lblCredentialsDomain.Enabled = radCredentialsCustom.Checked;
-            txtCredentialsUsername.Enabled = radCredentialsCustom.Checked;
-            txtCredentialsPassword.Enabled = radCredentialsCustom.Checked;
-            txtCredentialsDomain.Enabled = radCredentialsCustom.Checked;
-        }
+        
     }
 }
